rename: Renamed /example directory to /mock-data
