+++ mock-data/presentation-definition/mDL-example.json
